Fix `data-parent` in the AccordionGroup to refer to the Accordion
Not to be merged yet but to be discussed

Now the accordion unique ID is generated on the constructor of
Accordion. This allows to be able to pass it to the Container (which is
already created by this point).

Question is: should the unique ID be created on rendering? A this point,
is it too late to pass to the child? (is it already rendered)